ADD: specific service classes for contract transactions: - specific classes basically capture type of model; - base class alone should be avoided for use as service as of now; - added return types for base class for clarity; - `BaseContractTransactionService` captures `MongooseModel<ContractModel>` as common need.
